Consolidate error handling for custom apis
